Patch: Bring in new Live Display API
LiveDisplay is getting a proper API. This patch refactors
the settings to use it.

Also kill off the deprecated "gamma calibration" tunable
which nobody should be using at this point.

This Patch is does not include the removal of all translation resources.
Refer to the original CM patch, if you want those removed as well